# Major performance improvement in UDP processing by reusing buffer